refactor: add pin module
